openMP iteration variables can only be unsigned since OpenMP 3.0, and MSVC only supports OpenMP 2.0. Changed iteration variable to signed int.
git-svn-id: https://svn.code.sf.net/p/auber/code/tulip@4558 f5d4c5f9-c943-4916-8a57-bd8605bf961c

+++ plugins/metric/Eccentricity.cpp
@@ -137,7 +137,7 @@ bool EccentricityMetric::run() {
 #ifdef _OPENMP
 #pragma omp parallel for
 #endif
-  for (size_t ni = 0; ni < nbElem; ++ni) {
+  for (int ni = 0; ni < nbElem; ++ni) {
       if (stopfor) continue;
 #ifdef _OPENMP
     if (omp_get_thread_num() == 0) {
